Move test apps into test/support/apps
