+++ test/support/apps/rails/models/block6.rb
