+++ test/support/apps/rails/views/test/_message.html.erb
