+++ test/support/apps/rails/views/test/_syntax_error.html.erb
