+++ test/support/apps/rails/views/test/render_collection.html.erb
