+++ test/support/apps/rails/views/test/render_partial.html.erb
