+++ test/support/apps/rails/views/test/render_view.html.erb
