+++ test/support/apps/resque/boot.rb
@@ -1,0 +1,2 @@
+require_relative 'jobs/resque_error_job'
+require_relative 'jobs/resque_fast_job'
