+++ test/support/apps/resque/jobs/resque_error_job.rb
