+++ test/support/apps/resque/jobs/resque_fast_job.rb
