+++ test/test_helper.rb
@@ -27,41 +27,5 @@ WebMock.disable_net_connect!(
 
 Dir['test/support/*.rb'].each { |f| load(f) }
 
-# Boot background webservers to test against.
-require "./test/servers/rackapp_6511"
-
-case File.basename(ENV['BUNDLE_GEMFILE'])
-when /rails/
-  require './test/servers/rails_3205'
-when /grpc/
-  # Configure gRPC
-  require './test/servers/grpc_50051.rb'
-when /sidekiq/
-  # Hook into sidekiq to control the current mode
-  $sidekiq_mode = :client
-  class << Sidekiq
-    def server?
-      $sidekiq_mode == :server
-    end
-  end
-
-  ENV['REDIS_URL'] ||= 'redis://127.0.0.1:6379'
-
-  # Configure redis for sidekiq client
-  Sidekiq.configure_client do |config|
-    config.redis = { url: ENV['REDIS_URL'] }
-  end
-
-  # Configure redis for sidekiq worker
-  $sidekiq_mode = :server
-  ::Sidekiq.configure_server do |config|
-    config.redis = { url: ENV['REDIS_URL'] }
-  end
-  $sidekiq_mode = :client
-
-  require './test/servers/sidekiq/worker'
-end
-
 Minitest::Reporters.use! MiniTest::Reporters::SpecReporter.new
-
 Minitest::Test.include(Instana::TestHelpers)
